Initial starting code for direct objects (see commit description)
Created the object lists in lists.py for each type of object, so far being items, locations, and containers.  Created the Object class to store the type.  A list of Object wordtype objects was created for the parse_command function to eventually be used, along with adjectives, to build full descriptive names for each object.  This list of fullnames will be stored as direct objects on a per action basis.