Implemented a rough "FullObject" structure
(until a better name comes up) A FullObject class stores the fullname of an object including its adjectives.  For example, if the user entered "old cell key", the object is the key, and the adjectives are old and cell.  Since these words appeared in the users command, the parse_command function should have created word objects and stored them in the appropriate lists (adjective_objs, and object_objs).  The new get_full_objects function in the parse module takes those lists in, and connects the adjectives to the objects.  Then it builds a new FullObject object for every object in the object_objs list.  Once all of the new FullObject objects have been created and stored in a list, the list is returned.  The goal of this is to not have to worry about individual adjectives anymore, since they are now all connected to the objects.  This will allows us to build a system to check for direct objects in a users command.