Added a license file; changed saves directory
Added a license with the MIT licensing information.  Changed the saves directory depending on os.  If you are on Windows, the saves directory is places in $LOCALAPPDATA$\SevenStories\saves.  If you are using a POSIX interface, your saves directory will be in ~/.SevenStories/saves.  Retrieving the saves directory in QPython3 was cleaned slightly.