Implemented basic location system
Created the Location class, which for now only stores the location's name.  The GameMap class now creates a location with the name "First room".  This will be a temporary location just to place the player into, which is occurring as well.  The Player instance now stores that same Location object in the GameMap.  Also added an action "location" so that the user can check the player's location.